refactor(runtime): collapse RemoteMFE + AngularRemoteMFE into BaseRemoteMFE
RemoteMFE (React/rspack, 790 LOC) and AngularRemoteMFE (Angular/webpack, 796
LOC) shared ~700 lines of byte-identical logic — the atomic load lifecycle
with its telemetry checkpoints, the render flow, manifest introspection
(extractAvailableComponents / extractCapabilities), fetchContainer, and the
neutral capability handlers (refresh / authorizeAccess / health / describe /
schema / emit / updateControlPlaneState). That duplication was a standing
drift risk (two copies of the control-plane messaging, telemetry event names,
etc.).

Extracts all of it into a new framework-neutral BaseRemoteMFE. Each concrete
class now overrides only its framework surface — the ADR-036 model:
  • getSharedDependencies() — the MF shared-scope map (React vs Angular)
  • mountComponent()        — createRoot vs bootstrapApplication
  • unmount()               — release the React root vs destroy the ApplicationRef
AngularRemoteMFE also keeps resolveSelector() and its doQuery() override
("Query not supported"), which RemoteMFE deliberately does not have.

Result: the two framework files drop from 1,586 → 258 LOC (React 98,
Angular 160) — 84% off, well past the ~60% target — with the shared logic
single-sourced in base-remote-mfe.ts (761). Public API is unchanged: same
RemoteMFE / AngularRemoteMFE class names, exports, method signatures, and
behavior, so generated MFEs and existing tests are untouched.

boundary.test.ts: base-remote-mfe.ts added to NEUTRAL_FILES — it is now
machine-checked to import zero UI frameworks (the whole point of the split).

Coverage kept, not dropped: the shared logic's tests exercise it through the
concrete classes (base-remote-mfe 94.6% stmts / 83% branches). The React
adapter's mountComponent/unmount/getSharedDependencies were previously never
directly covered (the test harness stubs mountComponent); added parallel
React-adapter tests mirroring the Angular suite lift remote-mfe.ts from 16% →
96% stmts / 100% branches.

Verified end to end against examples/abc-kids:
  • React/rspack: flappy full Docker build OK (RemoteMFE hierarchy).
  • Angular/webpack: multiplication-quiz ng build passes (AngularRemoteMFE
    hierarchy); only the pre-existing unrelated .mesh/build:server step fails
    (this branch never touches that MFE).
Gates: lint 0 errors, typecheck clean, build clean (no schema drift),
boundary green, 96 suites / 1864 tests green, coverage gate passes.

Closes #237
