+++ src/runtime/__tests__/boundary.test.ts
@@ -26,13 +26,16 @@ const NEUTRAL_DIRS = ['packages/contracts/src'];
 
 // Individual neutral runtime files: boundary layer 4 (BaseMFE) plus the
 // framework-free control-plane core — LayoutManager, BaseControlPlane,
-// DaemonChannel, and the imperative-handle port. The contract shapes these
-// depend on now live in the scanned NEUTRAL_DIRS (@seans-mfe/contracts); the
-// inlined src/runtime/contracts.ts mirror was deleted in #236.
-// RemoteMFE / AngularRemoteMFE (layer 5) are intentionally excluded: they
-// produce the native handle and are allowed to import React / Angular.
+// DaemonChannel, the imperative-handle port, and BaseRemoteMFE (the neutral
+// Module Federation lifecycle shared by both framework adapters, ADR-036). The
+// contract shapes these depend on now live in the scanned NEUTRAL_DIRS
+// (@seans-mfe/contracts); the inlined src/runtime/contracts.ts mirror was
+// deleted in #236. RemoteMFE / AngularRemoteMFE (layer 5) are intentionally
+// excluded: they produce the native handle and are allowed to import React /
+// Angular.
 const NEUTRAL_FILES = [
   'src/runtime/base-mfe.ts',
+  'src/runtime/base-remote-mfe.ts',
   'src/runtime/context.ts',
   'src/runtime/graphql-ws-client.ts',
   'src/runtime/layout-manager.ts',
