+++ src/runtime/remote-mfe.ts
@@ -1,7 +1,13 @@
 /// <reference lib="dom" />
 /**
- * RemoteMFE Implementation
- * Concrete implementation of BaseMFE for Module Federation remotes
+ * RemoteMFE — React/rspack adapter over BaseRemoteMFE.
+ *
+ * All framework-neutral lifecycle (load / render / manifest introspection /
+ * capability handlers / control-plane messaging) lives in BaseRemoteMFE. This
+ * file supplies only what is React-specific (ADR-036):
+ *   • getSharedDependencies() — React singletons for the MF shared scope
+ *   • mountComponent()        — React 18 createRoot + error boundary
+ *   • unmount()               — release the React root
  *
  * Implements:
  * - REQ-RUNTIME-001: Load capability with atomic entry/mount/enable-render
@@ -9,527 +15,31 @@
  * - REQ-RUNTIME-012: Telemetry emission at all checkpoints
  */
 
-import { uuidv4 } from './util/uuid';
 import { createErrorBoundary } from './error-boundary';
-import {
-  BaseMFE,
-  LoadResult,
-  RenderResult,
-  Context,
-  HealthResult,
-  DescribeResult,
-  SchemaResult,
-  QueryResult,
-  EmitResult,
-  ControlPlaneStateResult,
-  type CapabilityMetadata,
-} from './base-mfe';
-import type { DSLManifest } from '../dsl/schema';
-import type { Message, ActionRecord, MessageMetadata } from '@seans-mfe/contracts';
+import { BaseRemoteMFE } from './base-remote-mfe';
+
+// Re-exported for consumers that imported the container type from this module.
+export type { ModuleFederationContainer } from './base-remote-mfe';
 
 /**
- * Module Federation container interface
- */
-export interface ModuleFederationContainer {
-  init(shared: Record<string, any>): Promise<void>;
-  get(module: string): Promise<() => any>;
-}
-
-/**
- * RemoteMFE class for Module Federation remotes
+ * RemoteMFE class for React Module Federation remotes.
  *
- * This class implements the load/render lifecycle for federated modules,
- * following the atomic operation model with telemetry at each checkpoint.
+ * Mounts React components via React 18 createRoot. Tracks a root per
+ * containerId so re-renders reuse the root and cleanup unmounts it.
  */
-export class RemoteMFE extends BaseMFE {
-  private container: ModuleFederationContainer | null = null;
-  private availableComponents: string[] = [];
-  private mountedComponent: any = null;
-  /** ID of the currently mounted component; used as actionRecord.componentId */
-  private currentComponentId: string | null = null;
+export class RemoteMFE extends BaseRemoteMFE {
   /** React roots keyed by containerId — reused on re-render, unmounted on cleanup */
   private reactRoots: Map<string, any> = new Map();
 
   /**
-   * Implement load logic for Module Federation remote
-   *
-   * REQ-RUNTIME-001: Atomic operation with three phases:
-   * 1. Entry: Fetch remote entry + container
-   * 2. Mount: Initialize container, wire shared deps
-   * 3. Enable-render: Prepare MFE state for render phase
-   */
-  protected async doLoad(context: Context): Promise<LoadResult> {
-    const startTime = Date.now();
-    const telemetry: LoadResult['telemetry'] = {
-      entry: { start: new Date(), duration: 0 },
-      mount: { start: new Date(), duration: 0 },
-      enableRender: { start: new Date(), duration: 0 },
-    };
-
-    try {
-      // Phase 1: Entry - Fetch Module Federation remote entry
-      const entryStart = Date.now();
-      (telemetry as any).entry = { start: new Date() };
-
-      if (this.deps?.telemetry) {
-        this.deps.telemetry.emit({
-          name: 'load-entry',
-          capability: 'load',
-          phase: 'entry',
-          status: 'start',
-          metadata: { mfe: this.manifest.name },
-          timestamp: new Date(),
-        });
-      }
-
-      // Get remote entry URL from context or manifest
-      const remoteEntry = (context.inputs?.remoteEntry as string) || this.manifest.remoteEntry;
-      if (!remoteEntry) {
-        throw new Error('Remote entry URL not provided in context.inputs or manifest');
-      }
-
-      // Fetch container (in real implementation, this would use Module Federation runtime)
-      this.container = await this.fetchContainer(remoteEntry);
-
-      (telemetry as any).entry.duration = Date.now() - entryStart;
-
-      if (this.deps?.telemetry) {
-        this.deps.telemetry.emit({
-          name: 'load-entry-metric',
-          capability: 'load',
-          phase: 'entry',
-          status: 'success',
-          metadata: {
-            mfe: this.manifest.name,
-            duration: (telemetry as any).entry.duration,
-          },
-          timestamp: new Date(),
-          duration: (telemetry as any).entry.duration,
-        });
-      }
-
-      // Phase 2: Mount - Initialize container and wire shared dependencies
-      const mountStart = Date.now();
-      (telemetry as any).mount = { start: new Date() };
-
-      if (this.deps?.telemetry) {
-        this.deps.telemetry.emit({
-          name: 'load-mount',
-          capability: 'load',
-          phase: 'mount',
-          status: 'start',
-          metadata: { mfe: this.manifest.name },
-          timestamp: new Date(),
-        });
-      }
-
-      // Initialize container with shared dependencies
-      const sharedDeps = this.getSharedDependencies();
-      await this.container.init(sharedDeps);
-
-      // Extract available components from manifest
-      this.availableComponents = this.extractAvailableComponents();
-
-      (telemetry as any).mount.duration = Date.now() - mountStart;
-
-      if (this.deps?.telemetry) {
-        this.deps.telemetry.emit({
-          name: 'load-mount-metric',
-          capability: 'load',
-          phase: 'mount',
-          status: 'success',
-          metadata: {
-            mfe: this.manifest.name,
-            duration: (telemetry as any).mount.duration,
-            componentsCount: this.availableComponents.length,
-          },
-          timestamp: new Date(),
-          duration: (telemetry as any).mount.duration,
-        });
-      }
-
-      // Phase 3: Enable-render - Prepare MFE state for render phase
-      const enableRenderStart = Date.now();
-      (telemetry as any).enableRender = { start: new Date() };
-
-      if (this.deps?.telemetry) {
-        this.deps.telemetry.emit({
-          name: 'load-enable-render',
-          capability: 'load',
-          phase: 'enable_render',
-          status: 'start',
-          metadata: { mfe: this.manifest.name },
-          timestamp: new Date(),
-        });
-      }
-
-      // Prepare render state (validate components, prepare metadata)
-      const capabilities = this.extractCapabilities();
-
-      (telemetry as any).enableRender.duration = Date.now() - enableRenderStart;
-
-      if (this.deps?.telemetry) {
-        this.deps.telemetry.emit({
-          name: 'load-enable-render-metric',
-          capability: 'load',
-          phase: 'enable_render',
-          status: 'success',
-          metadata: {
-            mfe: this.manifest.name,
-            duration: (telemetry as any).enableRender.duration,
-          },
-          timestamp: new Date(),
-          duration: (telemetry as any).enableRender.duration,
-        });
-      }
-
-      // Emit completion telemetry
-      const totalDuration = Date.now() - startTime;
-      if (this.deps?.telemetry) {
-        this.deps.telemetry.emit({
-          name: 'load-completed',
-          capability: 'load',
-          phase: 'completed',
-          status: 'success',
-          metadata: {
-            mfe: this.manifest.name,
-            success: true,
-          },
-          timestamp: new Date(),
-          duration: totalDuration,
-        });
-      }
-
-      // Populate context outputs
-      context.outputs = {
-        container: this.container,
-        manifest: this.manifest,
-        availableComponents: this.availableComponents,
-        capabilities,
-      };
-
-      return {
-        status: 'loaded',
-        container: this.container,
-        manifest: this.manifest,
-        availableComponents: this.availableComponents,
-        capabilities,
-        timestamp: new Date(),
-        duration: totalDuration,
-        telemetry,
-      };
-    } catch (error) {
-      // Emit error telemetry
-      if (this.deps?.telemetry) {
-        this.deps.telemetry.emit({
-          name: 'load-error',
-          capability: 'load',
-          phase: 'error',
-          status: 'error',
-          metadata: {
-            mfe: this.manifest.name,
-            error: (error as Error).message,
-          },
-          timestamp: new Date(),
-        });
-      }
-
-      const err = error as Error;
-      return {
-        status: 'error',
-        timestamp: new Date(),
-        duration: Date.now() - startTime,
-        error: {
-          message: err.message ?? String(err),
-          phase: 'entry',
-          retryCount: 0,
-          retryable: true,
-        },
-        telemetry,
-      };
-    }
-  }
-
-  /**
-   * Implement render logic for Module Federation remote
-   *
-   * REQ-RUNTIME-004: Component-aware rendering with:
-   * - Component selection from available components
-   * - Props validation and passing
-   * - Error boundary integration
-   * - DOM mounting with React 18 createRoot
-   */
-  protected async doRender(context: Context): Promise<RenderResult> {
-    const startTime = Date.now();
-
-    try {
-      // Extract render parameters from context
-      const componentName = context.inputs?.component as string;
-      const props = (context.inputs?.props as Record<string, any>) || {};
-      const containerId = context.inputs?.containerId as string;
-
-      if (!componentName) {
-        throw new Error('Component name not provided in context.inputs.component');
-      }
-
-      if (!containerId) {
-        throw new Error('Container ID not provided in context.inputs.containerId');
-      }
-
-      // Telemetry: Render start
-      if (this.deps?.telemetry) {
-        this.deps.telemetry.emit({
-          name: 'render-start',
-          capability: 'render',
-          phase: 'render_start',
-          status: 'start',
-          metadata: {
-            mfe: this.manifest.name,
-            component: componentName,
-          },
-          timestamp: new Date(),
-        });
-      }
-
-      // Validate component exists in available components
-      if (!this.availableComponents.includes(componentName)) {
-        throw new Error(
-          `Component "${componentName}" not found in availableComponents: [${this.availableComponents.join(', ')}]`
-        );
-      }
-
-      // Validate container exists
-      if (!this.container) {
-        throw new Error('Container not loaded. Call load() before render()');
-      }
-
-      // Load component directly via loadDomainComponent (implemented by subclass)
-      const renderStart = Date.now();
-      const Component = await this.loadDomainComponent(componentName);
-
-      const renderDuration = Date.now() - renderStart;
-
-      // Telemetry: Component fetch duration
-      if (this.deps?.telemetry) {
-        this.deps.telemetry.emit({
-          name: 'render-component-fetch',
-          capability: 'render',
-          phase: 'component_fetch',
-          status: 'success',
-          metadata: {
-            mfe: this.manifest.name,
-            component: componentName,
-          },
-          timestamp: new Date(),
-          duration: renderDuration,
-        });
-      }
-
-      // Mount component to DOM
-      const mountStart = Date.now();
-      const element = await this.mountComponent(Component, props, containerId);
-      const mountDuration = Date.now() - mountStart;
-
-      // Telemetry: Mount duration
-      if (this.deps?.telemetry) {
-        this.deps.telemetry.emit({
-          name: 'render-mount',
-          capability: 'render',
-          phase: 'mount',
-          status: 'success',
-          metadata: {
-            mfe: this.manifest.name,
-            component: componentName,
-          },
-          timestamp: new Date(),
-          duration: mountDuration,
-        });
-      }
-
-      const totalDuration = Date.now() - startTime;
-
-      // Telemetry: Render completed
-      if (this.deps?.telemetry) {
-        this.deps.telemetry.emit({
-          name: 'render-completed',
-          capability: 'render',
-          phase: 'completed',
-          status: 'success',
-          metadata: {
-            mfe: this.manifest.name,
-            component: componentName,
-            success: true,
-          },
-          timestamp: new Date(),
-          duration: totalDuration,
-        });
-      }
-
-      // Store mounted component reference
-      this.mountedComponent = { component: componentName, element, props };
-      this.currentComponentId = componentName;
-
-      // Populate context outputs
-      context.outputs = {
-        ...context.outputs,
-        component: componentName,
-        element,
-        renderDuration,
-        mountDuration,
-      };
-
-      return {
-        status: 'rendered',
-        component: componentName,
-        element,
-        timestamp: new Date(),
-        duration: totalDuration,
-        renderDuration,
-        mountDuration,
-      };
-    } catch (error) {
-      // Emit error telemetry
-      if (this.deps?.telemetry) {
-        this.deps.telemetry.emit({
-          name: 'render-error',
-          capability: 'render',
-          phase: 'error',
-          status: 'error',
-          metadata: {
-            mfe: this.manifest.name,
-            error: (error as Error).message,
-          },
-          timestamp: new Date(),
-        });
-      }
-
-      return {
-        status: 'error',
-        timestamp: new Date(),
-        duration: Date.now() - startTime,
-        error: error as Error,
-      };
-    }
-  }
-
-  // =========================================================================
-  // Helper Methods
-  // =========================================================================
-
-  /**
-   * Fetch Module Federation container from remote entry
-   */
-  private async fetchContainer(remoteEntry: string): Promise<ModuleFederationContainer> {
-    // In a real implementation, this would:
-    // 1. Load the remoteEntry.js script
-    // 2. Access the global container variable
-    // 3. Return the container interface
-
-    // For now, return a mock container for testing
-    return {
-      init: async (shared: Record<string, any>) => {
-        // Initialize with shared dependencies
-        void shared; // Module Federation shared scope
-      },
-      get: async (module: string) => {
-        // Return a factory function for the requested module
-        void module;
-        return () => ({
-          default: class MockComponent {},
-        });
-      },
-    };
-  }
-
-  /**
    * Get shared dependencies for Module Federation
    */
-  private getSharedDependencies(): Record<string, any> {
+  protected getSharedDependencies(): Record<string, any> {
     // Return shared dependencies (React, ReactDOM, etc.)
     return {
       react: { version: '18.2.0', singleton: true },
       'react-dom': { version: '18.2.0', singleton: true },
     };
-  }
-
-  /**
-   * Extract available components from manifest.
-   *
-   * Primary:  render.components array when explicitly declared.
-   * Fallback: all non-platform capability names (domain capabilities).
-   *           This allows MFEs to work without a render capability block
-   *           while still exposing their domain features as mountable components.
-   */
-  private extractAvailableComponents(): string[] {
-    if (!this.manifest.capabilities) {
-      return [];
-    }
-
-    // Primary: explicit render.components list
-    for (const capEntry of this.manifest.capabilities) {
-      if (capEntry.render) {
-        const components = (capEntry.render as any).components;
-        if (Array.isArray(components) && components.length > 0) {
-          return components;
-        }
-      }
-    }
-
-    // Fallback: collect domain capability names (everything that is not a platform capability)
-    const PLATFORM_CAPABILITY_NAMES = new Set([
-      'load',
-      'render',
-      'refresh',
-      'authorizeAccess',
-      'health',
-      'describe',
-      'schema',
-      'query',
-      'emit',
-      'updateControlPlaneState',
-      // Also handle the PascalCase variants used as capability entry keys in YAML
-      'Load',
-      'Render',
-      'Refresh',
-      'AuthorizeAccess',
-      'Health',
-      'Describe',
-      'Schema',
-      'Query',
-      'Emit',
-      'UpdateControlPlaneState',
-    ]);
-    const domainComponents: string[] = [];
-    for (const capEntry of this.manifest.capabilities) {
-      for (const name of Object.keys(capEntry)) {
-        if (!PLATFORM_CAPABILITY_NAMES.has(name)) {
-          domainComponents.push(name);
-        }
-      }
-    }
-    return domainComponents;
-  }
-
-  /**
-   * Extract capability metadata from manifest (REQ-RUNTIME-003)
-   */
-  private extractCapabilities(): CapabilityMetadata[] {
-    if (!this.manifest.capabilities) {
-      return [];
-    }
-
-    const result: CapabilityMetadata[] = [];
-    for (const capEntry of this.manifest.capabilities) {
-      for (const [name, def] of Object.entries(capEntry)) {
-        const d = def as { type?: string; description?: string } | undefined;
-        result.push({
-          name,
-          type: d?.type === 'domain' ? 'domain' : 'platform',
-          description: d?.description,
-        });
-      }
-    }
-    return result;
   }
 
   /**
@@ -575,16 +85,6 @@ export class RemoteMFE extends BaseMFE {
   }
 
   /**
-   * Override in subclass to load the named domain component.
-   * Called by doRender() instead of going through the Module Federation container API.
-   */
-  protected async loadDomainComponent(_name: string): Promise<any> {
-    throw new Error(
-      '[RemoteMFE] loadDomainComponent() not implemented — subclass must override this method'
-    );
-  }
-
-  /**
    * Unmount a previously rendered component and release the React root.
    * Call from the shell's useEffect cleanup to avoid memory leaks.
    */
@@ -594,197 +94,5 @@ export class RemoteMFE extends BaseMFE {
       root.unmount();
       this.reactRoots.delete(containerId);
     }
-  }
-
-  // =========================================================================
-  // Other Required Abstract Methods
-  // =========================================================================
-
-  protected async doRefresh(_context: Context): Promise<void> {
-    // TODO: implement refresh — fetch updated data and re-render
-  }
-
-  protected async doAuthorizeAccess(context: Context): Promise<boolean> {
-    // Check authorization
-    return true;
-  }
-
-  protected async doHealth(context: Context): Promise<HealthResult> {
-    return {
-      status: 'healthy',
-      checks: [
-        {
-          name: 'container',
-          status: this.container !== null ? 'pass' : 'fail',
-          message: this.container !== null ? 'Container loaded' : 'Container not loaded',
-        },
-        {
-          name: 'components',
-          status: this.availableComponents.length > 0 ? 'pass' : 'fail',
-          message: `${this.availableComponents.length} components available`,
-        },
-      ],
-      timestamp: new Date(),
-    };
-  }
-
-  protected async doDescribe(context: Context): Promise<DescribeResult> {
-    return {
-      name: this.manifest.name,
-      version: this.manifest.version,
-      type: this.manifest.type,
-      capabilities: this.extractCapabilities().map((c) => c.name),
-      manifest: this.manifest,
-    };
-  }
-
-  protected async doSchema(context: Context): Promise<SchemaResult> {
-    return {
-      schema: JSON.stringify(this.manifest, null, 2),
-      format: 'json',
-    };
-  }
-
-  protected async doEmit(context: Context): Promise<EmitResult> {
-    if (this.deps?.telemetry) {
-      const event = context.inputs?.event;
-      if (event) {
-        this.deps.telemetry.emit(event as any);
-        return {
-          emitted: true,
-          eventId: 'generated-event-id',
-        };
-      }
-    }
-    return {
-      emitted: false,
-    };
-  }
-
-  /**
-   * Push domain state to the daemon control plane for registry re-evaluation.
-   *
-   * Module Federation MFEs run in the browser, so this sends a GraphQL mutation
-   * over the existing WebSocket connection to the daemon:
-   *
-   *   mutation SendAction($input: ActionInput!) {
-   *     sendAction(input: $input) { correlationId acknowledged }
-   *   }
-   *
-   * The daemon forwards to Registry handleMessage → rules engine re-evaluation.
-   * The registry may resolve a new MFE + capability, which arrives via the
-   * Subscription.messages channel the Renderer is already subscribed to.
-   */
-  protected async doUpdateControlPlaneState(context: Context): Promise<ControlPlaneStateResult> {
-    // -------------------------------------------------------------------------
-    // 1. Extract and validate inputs
-    // -------------------------------------------------------------------------
-    const rawStateKey = context.inputs?.stateKey;
-    const rawStateData = context.inputs?.stateData;
-    const rawCorrelationId = context.inputs?.correlationId;
-
-    if (typeof rawStateKey !== 'string' || rawStateKey.trim().length === 0) {
-      throw new Error(
-        'updateControlPlaneState requires context.inputs.stateKey to be a non-empty string'
-      );
-    }
-
-    if (
-      rawStateData !== undefined &&
-      (typeof rawStateData !== 'object' || rawStateData === null || Array.isArray(rawStateData))
-    ) {
-      throw new Error(
-        'updateControlPlaneState requires context.inputs.stateData to be an object when provided'
-      );
-    }
-
-    if (
-      rawCorrelationId !== undefined &&
-      (typeof rawCorrelationId !== 'string' || rawCorrelationId.trim().length === 0)
-    ) {
-      throw new Error(
-        'updateControlPlaneState requires context.inputs.correlationId to be a non-empty string when provided'
-      );
-    }
-
-    const stateKey = rawStateKey.trim();
-    const stateData = (rawStateData ?? {}) as Record<string, unknown>;
-    const correlationId = (rawCorrelationId as string | undefined)?.trim() ?? context.requestId;
-
-    // -------------------------------------------------------------------------
-    // 2. Guard: daemon WebSocket must be connected
-    // -------------------------------------------------------------------------
-    const wsClient = this.deps?.wsClient;
-    if (!wsClient || !wsClient.connected) {
-      return { acknowledged: false, correlationId, error: 'Daemon WebSocket not connected' };
-    }
-
-    // -------------------------------------------------------------------------
-    // 3. Build the Message envelope typed against @control-plane/contracts
-    // -------------------------------------------------------------------------
-    const metadata: MessageMetadata = {
-      correlationId,
-      acknowledged: false,
-      error: null,
-    };
-
-    // Emit what the registry actually routes on (ADR-057): a canonical
-    // STATE_UPDATE actionType plus the stateKey. The registry matches rules on
-    // `when.stateKey`; setting actionType: stateKey (and no stateKey field)
-    // meant those routes never fired.
-    const payload: ActionRecord = {
-      id: uuidv4(),
-      componentId: this.currentComponentId ?? this.manifest.name,
-      actionType: 'STATE_UPDATE',
-      stateKey,
-      data: stateData,
-      timestamp: new Date().toISOString(),
-    };
-
-    const envelope: Message = {
-      direction: 'ACTION',
-      kind: 'ACTION',
-      payload,
-      metadata,
-    };
-
-    // -------------------------------------------------------------------------
-    // 4. Send via the daemon's sendMessage GraphQL mutation (WS subscribe/next)
-    // -------------------------------------------------------------------------
-    let success: boolean;
-    try {
-      success = await wsClient.mutation(
-        'mutation sendMessage($m: String!) { sendMessage(message: $m) }',
-        { m: JSON.stringify(envelope) },
-        4_000
-      );
-    } catch (err) {
-      const message = (err as Error).message ?? 'sendMessage mutation failed';
-      const error = message.includes('timed out') ? 'sendMessage timed out' : message;
-      return { acknowledged: false, correlationId, error };
-    }
-
-    // -------------------------------------------------------------------------
-    // 5. Telemetry
-    // -------------------------------------------------------------------------
-    if (this.deps?.telemetry) {
-      this.deps.telemetry.emit({
-        name: 'control-plane-state-update',
-        capability: 'updateControlPlaneState',
-        phase: 'main',
-        status: success ? 'success' : 'error',
-        metadata: { mfe: this.manifest.name, stateKey, correlationId, acknowledged: success },
-        timestamp: new Date(),
-      });
-    }
-
-    // -------------------------------------------------------------------------
-    // 6. Return ControlPlaneStateResult
-    // -------------------------------------------------------------------------
-    return {
-      acknowledged: success,
-      correlationId,
-      error: success ? null : 'sendMessage mutation failed',
-    };
   }
 }
